FilterSurroundingCode moved to code-processing
