+++ src/tools/create/inline_code_completion.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import { GeminiRepository } from '../../repository/gemini-repository';
 import { logError, logEvent } from '../../utilities/telemetry-reporter';
 import path = require('path');
-import { extractDartCode } from '../../utilities/code-processing';
+import { extractDartCode, filterSurroundingCode } from '../../utilities/code-processing';
 import { ContextualCodeProvider } from '../../utilities/contextual-code';
 import { ILspAnalyzer } from '../../shared/types/LspAnalyzer';
 import { dartCodeExtensionIdentifier } from '../../shared/types/constants';
@@ -162,75 +162,3 @@ function getDartAnalyser() { // This could be in a wider scope.
     const analyzer: ILspAnalyzer = dartExt?.exports._privateApi.analyzer;
     return analyzer;
 }
-
-
-
-export function filterSurroundingCode(orignalContent: string, codeCompletion: string, splitLineNumberStart: number, splitLineNumberEnd: number = splitLineNumberStart): string {
-    const orginalContentLines = orignalContent.split('\n');
-    let codeCompletionLines = codeCompletion.split('\n');
-
-    const preInsertLines = orginalContentLines.slice(0, splitLineNumberStart);
-    const afterInsertLines = orginalContentLines.slice(splitLineNumberEnd + 1);
-
-    const codeCompletionStartLine = removeWhitespaces(codeCompletionLines[0]);
-    for (let i = 0; i < preInsertLines.length; i++) { //5
-        if (codeCompletionLines.length < preInsertLines.length - i) {
-            continue; // surrounding line is out of code completion range.
-        }
-        //find the first line from the top of the document that matches the starting line of code completion.
-        const existingLine = removeWhitespaces(preInsertLines[i]);
-        if (codeCompletionStartLine === existingLine) {
-            let fullMatch = true;
-            // all the below lines in the document should also match with the code completion
-            for (let j = 1; j < preInsertLines.length - i; j++) {
-                const followingCodeCompletionLine = removeWhitespaces(codeCompletionLines[j]);
-                const followingExistingLine = removeWhitespaces(preInsertLines[i + j]);
-                if (followingCodeCompletionLine !== followingExistingLine) {
-                    fullMatch = false;
-                    break;
-                }
-            }
-            if (fullMatch) {
-                codeCompletionLines = codeCompletionLines.slice(preInsertLines.length - i);
-                break;
-            }
-        }
-    }
-
-    // Cleanup logic for after lines
-    if(codeCompletionLines.length>0){ //safe for all code completion lines removed in last step resulting into error in codeCompletionLines[codeCompletionLines.length - 1]
-        const codeCompletionEndLine = removeWhitespaces(codeCompletionLines[codeCompletionLines.length - 1]);
-        for (let i = afterInsertLines.length; i > 0; i--) {
-            if (codeCompletionLines.length < i) {
-                continue; // surrounding line is out of code completion range.
-            }
-            //find the last line of the doc that matches with the last line of code completion
-            const existingLine = removeWhitespaces(afterInsertLines[i - 1]);
-            if (codeCompletionEndLine === existingLine) {
-                let fullMatch = true;
-                // make sure all the lines from last line in doc to the line after cursor are available in code compleiton
-                for (let j = 1; j < i; j++) {
-                    const previousCodeCompletionLine = removeWhitespaces(codeCompletionLines[codeCompletionLines.length - 1 - j]);
-                    const previousExistingLine = removeWhitespaces(afterInsertLines[i - 1 - j]);
-                    if (previousCodeCompletionLine !== previousExistingLine) {
-                        fullMatch = false;
-                        break;
-                    }
-                }
-                if (fullMatch) {
-                    codeCompletionLines = codeCompletionLines.slice(0, codeCompletionLines.length - i);
-                    break;
-                }
-            }
-        }
-    }
-    
-    // Join the cleaned up code completion lines with the original content lines
-    const result = codeCompletionLines.join('\n');
-    return result;
-}
-
-function removeWhitespaces(line: string): string {
-    return line.replace(/\s/g, "");
-}
-
